+++ Backend/test-api.http
@@ -1,19 +1,23 @@
 ###
+
 POST localhost:8000/api/autoUsate/
 Accept: application/json
 Content-Type: application/json
 
 {
-  "modello": "Fiat Punto",
-  "marca": "Fiat",
-  "altezza": 1500,
-  "lunghezza": 1500,
-  "larghezza": 1500,
-  "peso": 25,
-  "volume_bagagliaio": 12,
-  "km_percorsi": 150000,
-  "anno_immatricolazione": "2023-01-01",
-  "targa": "AA123BB"
+  "utente": 1,
+  "auto": {
+    "modello": "Fiat Punto",
+    "marca": "Fiat",
+    "altezza": 1500,
+    "lunghezza": 1500,
+    "larghezza": 1500,
+    "peso": 25,
+    "volume_bagagliaio": 12,
+    "km_percorsi": 150000,
+    "anno_immatricolazione": "2023-01-01",
+    "targa": "AA123BB"
+  }
 }
 
 ###
@@ -57,7 +61,7 @@ Content-Type: application/json
   },
   "optional": [
     1,
-    5
+    8
   ]
 }
